adds a user store
This store provides ways for creating and mutating a user.
When a user is created, they get an email with a link to reset their email.
Updating a user only updates their permissions. If a user wants to reset their
email, we need to implement a method that will trigger the SendPasswordResetEmail.